fixes + param optimisation for all 8 algos

+++ climate-envs/climate_envs/envs/simple_climate_bias_correction.py
@@ -140,7 +140,7 @@ class SimpleClimateBiasCorrectionEnv(gym.Env):
             np.array: The current observation.
         """
         temperature = self.state[0]
-        return np.array([temperature])
+        return np.array([temperature], dtype=np.float32)
 
     def render(self):
         if self.render_mode == "rgb_array":
